pyproject.toml
[tool.poetry]
name = "m3uedit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyaml = "^23.12.0"
m3u-ipytv = "^0.2.7"
pydantic = "^2.6.4"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
pyright = "^1.1.354"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
